match: nb_sess cmd param -> short (matches decomp/CONCAT22 return); 74.19%
Correctness refinement to the nb_sess signature (the retail movsx-reads the command + the return
CONCAT22s its sign bit). The remaining ~26% is deeper switch-dispatch / return-format / NCB-field
detail to iterate. The 8-static callback web is in and building.

+++ src/SOURCE/netwin.cpp
@@ -381,7 +381,7 @@ static void __fastcall nb_format_name(char *src, unsigned char *dst)
 }
 
 VA(0x004a726a, 0x4cd)
-extern "C" int __cdecl nb_sess(int cmd, int session, char *name, int detach)
+extern "C" int __cdecl nb_sess(short cmd, int session, char *name, int detach)
 {
     unsigned short result;
     NCB ncb;
